+++ frontend/src/components/atoms/Features/AboutSidebar.tsx
